Rename test_entries to entries
